Fix issue with VPN not being active due to IPv6
In order to activate the change regarding the usage of IPv6, the module has to be rebooted.

However, VPNs disconnect when the connections stops. Thus, the order has been changed.

Also there is no useless attempt to turn off the VPN when it is not even on.

Also there are now somewhat helpful output messages.