fix VPN not connecting properly to custom VPN
I am stupid and I used the wrong param. Obviously it didn't work but all I could test was the one I already knew worked. And conveniently, I just put it in the code as well. For good measure? Don't ask me - I don't know either.

Also I changed it so that the IPv6-Traffic will not get deactivated to profile UPB (Uni-VPN Paderborn). Might make sense to do it the other way around, i.e. only activate it for certain ones. Or maybe do it as an additional param?